Author: Rex Kerr
I finished a first rough draft of a MCell3 quick tutorial and reference
guide.  I'm not sure it's any good as either (since it's kind of a strange
mishmash of the two), but here it is.

It's created in LyX, which is essentially a visual Latex editor with a few
modifications/extensions.

Additions and/or suggestions are welcome.