Author: Rex Kerr
Added a new type of count statement, and fixed a number of bugs in the old
count statements for hit types (mostly uninitialized variables).

I'm still not sure if this is working perfectly, but it shouldn't have
broken anything we were previously using.

The new count type is CONCENTRATION, and is currently implemented for
regions only, and is documented in the QRG.  It uses the number of impacts
to determine the effective concentration at a surface.

I also changed the code in the mcell.c main loop back to a more compact
form with for loops, since the uncompact form wasn't achieving anything with
its lack of compactness.

I might have removed some debug messages, or I might have added more.  I'm
not sure at this point.