Author: Rex Kerr
The count code wasn't properly evaluating expressions after the first round
of output to files.  Now it does (and the code is marginally simpler, too).
This was easy to notice with the new COUNT[molecule,region,CONCENTRATION]
command since internally it uses an expression.  (Likewise with ALL_HITS and
ALL_CROSSINGS, but those hadn't been used enough yet to notice.)