Author: Rex Kerr
Fixed a memory leak (inasmuch as having no roof counts as a "leak" in the
roof) where some grid molecules were not being deallocated after they were
destroyed.  They also appear to not be in the scheduler (and I'd thought I
was sticking everything in there), which could make it tricky to find them
for visualization purposes, but that's a project for another day.