Author: Rex Kerr
I added the ability to send multiple columns of output to a single file.
The syntax is, not surprisingly,

   { COUNT[whatever] , COUNT[somethingelse] , ... } => "my_favorite.file"

To do this, I added a "next_column" pointer in output_item and iterate over
that for each output_item in a list.

Also, we so far have only been supporting an index_type of TIME_STAMP_VAL.
If this is all we need, we can remove the index_type variable from the
output_evaluator struct.  (I left it alone for now, just in case, but having
multiple index types for output doesn't really make sense when you've got
more than one count statement per line.)

Note also that this may have made output a little less efficient, since I'm
calling fprintf a few more times.  Hopefully it has decent internal
buffering or we should buffer the output to strings before sending it to
files.

Finally, I updated the QRG to reflect this new feature.