Author: Rex Kerr
Counting reactions in the world was broken because reaction pathways got
reallocated after their counters were set, leading to dangling pointers and
all sorts of fun stuff like that.

To fix this, pathways now keep track of which output_evaluators are looking
at them via a struct pathway_count_request (linked list per pathway).  When
reactions are prepared and the pathways are reallocated from a linked list
into an array, the output_evaluator pointer is updated accordingly.  The
code now appears to work for counting reactions in the world.

Counting reactions in other places is known to not work yet.