Author: Rex Kerr (and Tom Bartol)
Tom and I added a new random number generator.  It is somewhat faster than
the previous generator, and much less sensitive to picking magically good
seed numbers.  If you compile with this new random number generator
(isaac64), you can choose seed values up to 2^31-1.  Otherwise you're still
limited to 3000.  The new generator is on by default.  The nmj_recon1 model
comes out the same (to within stochastic noise) with the new generator, so
it should be working fine.

Also, I generalized the way that we deal with random numbers throughout the
code.  Inside rng.h, all the internals of the different random number
generators are #defined away, leaving you with a struct rng_state that
stores the state of your generator, and functions rng_uint, rng_dbl, and
rng_gauss which return 32 bit integers, double values in the interval [0,1),
and double values in a Gaussian distribution with mean 0 and standard
deviation 1, respectively.  Each of these functions takes a struct rng_state
as its only argument.

Just in case you're wondering, the reason this is being adding is so we can
more easily test cases where it would be helpful to average over more than
3000 cases (without doing silly things like replicating the world a bunch of
times).