Author: Rex Kerr
If the user asks to count something on an object with no countable geometry,
the code allowed it, and then broke when trying to write out files.  It was
easier to fix it so that the output code omits warnings instead of errors,
and this is in general a better policy since in an error condition we want
to preserve as much data as possible while telling the user what's gone
wrong.

So this case now generates lots of warning messages.  At some point, we
should also probably make it so that you cannot count on release objects and
collections thereof.