Author: Rex Kerr (with help from Tom)
We fixed the parser so it is now a parse-time error to try to count
molecules inside a metaobject filled exclusively with release sites (which
can't be counted on).

We also agreed to adopt the policy that functions called by mdlparse.y will
output an error message without a line number and return an error code to
mdlparse.y, which would then output a more general error message including
the line number.  (If the called function cannot return for some reason, it
probably ought to spit out the line number right then.)  The function
"mdlerror_nested" allows the functions in mdlparse_util.c to print out an
error without giving the line number.