Author: Rex Kerr
Fixed a bug in the handling of max_time in diffuse.c.  It wasn't calling the
right code for diffusing particles, so I rearranged things.

Also, fixed a division-by-zero error in timeof_unimolecular in react_cond.c
(it would crash if it got a zero from the random number generator).