Fixed an off-by-one error that led to one extra iteration being run at the end of a simulation.
Also fixed an error in handling world counts of reaction output--it now
should avoid crashing.

Added a line for -march=athlon-xp in the makedefs.  This is the best setting
to use for opterons--and it runs very well on 32-bit athlons and on p4s,
also, so I use it for everything now just for convenience.  I've left the
-march=pentium4 as the default, though.

Also fiddled with stuff in diffuse.c and wall_util.c for debugging purposes,
but I don't think I ended up changing anything.