Author: Rex Kerr
I've finally written exact_disk, which should give the exact fractional area
accessible for the interaction disk of a moving molecule with an arbitrary
number of arbitrarily intersecting surfaces.  The function has not been
carefully debugged yet (it seems to work okay for normal cases, however),
but I'm committing it because it's ~800 lines of carefully optimized code
and I'd rather use CVS to protect it from inadvertant destruction.

Preliminary performance tests indicate that it is approximately the same
speed as the old estimate_disk, although this is expected to be a function
of the number of walls in the partition (fewer walls/more occlusions will
probably favor estimate_disk).

For now, I've left estimate_disk as the function that is called by
diffuse_3D.  The functions have identical usage, however, so all that is
required to switch is to change "estimate" to "exact".