Author: Rex Kerr
Added documentation for CUSTOM_SPACE_STEP, OUTPUT_BUFFER_SIZE, RAND_UNIFORM,
and RAND_GAUSSIAN.

Also added basic sanity-checking to the value supplied to OUTPUT_BUFFER_SIZE
(i.e. >0 and fits in an unsigned int).  Making this number too large is
still a great way for a user to generate an out-of-memory error, however.