Author: Rex Kerr
Fixed another x/z transposition in exact_disk.  I've also moved exact_disk
to be the default code used to look for disk occlusion.

Also changed almost everything that expects iterations to be an integer (4
bytes) so that it uses a long long (8 bytes) instead.  The one exception is
output for DX; I'm not sure what is necessary there, so I've left it as an
int.  I've run simulations of 8G iterations without problem (but without viz
output).

Also reordered the code that handles custom space/time steps to avoid errors
when the diffusion constant is set to zero.

Finally, I removed the path_length variable from molecules and put birthday
in its place (which says when the molecule was first created).  This is then
used to gradually ramp up the molecule's timestep after its birth until it
reaches its full value (if the full value is greater than one world
timestep); doing this gives more accurate rebinding calculations.