Author: Rex Kerr
Added code to collect data on statistical errors made by MCell3.  In
particular, each species keeps track of its average lifetime and each
reaction keeps track of the number of times the probability has exceeded
one.  At the end of the simulation, a warning is printed if the mean
lifetime is less than 50 timesteps (which could indicate a significant
rounding error), or if more than 0.1% of the reactions have been skipped
because the probability was too high.

Also, I turned off reporting the number of times each reaction has occurred.
This can be enabled again by setting the correct block of code in mcell.c to
#if 1.  (Line 241 at the moment.)