Author: Rex Kerr
Changed default effector grid density to 10000 (init.c).

Modified set_partitions so that it will not create partitions smaller than
the interaction diameter (interaction radius * 2).  If the user specifies
smaller partitions, the program quits with an error.  Default partitions are
chosen to be larger.  The code isn't heavily tested, but it seems to be
working on my sample cases.

About half of the code for releases on regions is in.  The old syntax still
works, so there shouldn't be any difference in the way simulations currently
run.  There is new syntax for release sites which is

name RELEASE_SITE { SHAPE=... /*location, molecule, etc.*/ }

SHAPE can be set to SPHERICAL, CUBIC, or SPHERICAL_SHELL.  SHAPE can also be
set to the name of a region or a bunch of regions composed with set
operations + (union), - (subtraction), and * (intersection).  So you can
release on anything you can specify, on or in one or multiple objects.

The region part doesn't work yet (if you try it, it'll parse and then
probably give weird errors or do nothing).  But there seems no harm in
committing the half-complete stuff now.  When it's done, it will be an error
to release a 3D molecule inside a region that is not fully closed.

Also, util.c has some new functions for dealing with linked lists and arrays
of void pointers.  I haven't tested them; that will happen when the region
release code is finished.