Author: Rex Kerr
Release-site-on-region functionality is now in.  There may still be some
bugs; I've only given it a cursory testing with boxes.  The cursory testing
looks good.

I still need to add more error messages and I probably should add a test
that bails out of releases in 3D regions if it can't ever find a place to
put a molecule.  (It is possible to have exactly zero valid volume, and
unlike with 2D regions, it takes a while to realize this.)

I also need to comment the code.  Some of it is nontrivial, as one might
expect from a nontrivial operation like adding molecules to arbitrary
combinations of regions of surfaces and/or 3D space.

There were quite a few bugfixes for the stuff I'd committed working up to
this release, hence the number of files changed.

Note: there is a performance penalty for releasing on regions--you have to
use the counting code to keep track of whether molecules are in a region or
not, and it's harder to figure out where to put stuff than with normal
release commands or initialization with a particular surface density of
receptors.  I don't think it's too bad (I tried to make the code as
efficient as possible without spending too many weeks coding it), but the
performance should be tested.