Author: Rex Kerr
Release regions now work if defined as non-instanced objects.  However, in
order for them to know which regions they're looking at, they must be placed
inside a meta-object that also contains the region that they refer to.
Alternatively, the release site can refer to any already-instantiated
regions anywhere in space.

It's an error to try to instantiate just the release site without the full
object that contains all the pieces that the site needs to refer to.

Also fixed a bug in releasing on surface regions.

I again haven't tested this extensively, but it seems to work.