Author: Rex Kerr
Fixed non-region release sites so that they scale with all the normal
transformations.  Also fixed a bug in regions copying--they weren't getting
their flags copied properly.  (I think this was a bug....)

Also, if you specify CONCENTRATION= instead of NUMBER_TO_RELEASE= you can
set a concentration for the region (it'll do the appropriate calculation for
you).

As it stands now, if you use a scaling transformation to change the size of
the region, it'll spread out or compress that concentration for normal
shapes, but use the stated concentration for region release sites.  We
probably ought to decide on one policy and make them both work that way.
(It works the way it does now because this is the easiest.)