Author: Rex Kerr
Updated the QRG to reflect the fact that molecule names and surface names
must be unique within the entire simulation (not just MDL file) and that
they must be distinct from each other.  Also mentioned that a surface name can
be one reaction partner in a reaction, which explains why the names must be
different from each other.