Author: Rex Kerr
Added the ability for grid molecules to land at arbitrary locations on
tiles.  It seems to be working fine--gives similar results as centered
location with nmj_recon_1.  I haven't tested to see if there's a significant
reduction with a large sample size.  Some reduction may be expected, because
if a released molecule isn't centered over a receptor, it's less likely to
rebind.

The feature is off by default for now.  Use
  CENTER_MOLECULES_ON_GRID = FALSE
in the MDL file to turn it on (TRUE is the default).

Also, added the option
  ACCURATE_3D_REACTIONS = FALSE
to turn off expanded list calculations.  The default is TRUE for now
(meaning the expanded list is used by default).

I also removed the #ifdef stuff for the expanded collision list since it's
now controlled by the world->use_expanded_list variable.

Note: even though there are no 3D reactions in the nmj_recon_1 simulation,
it takes about 15% longer with ACCURATE_3D_REACTIONS=TRUE.  A good portion
of this is during initialization, as adding walls to subvolumes takes a lot
longer.  There's some additional slowdown during runtime, I think, since
more walls are traced against.  We probably should switch this to FALSE if
there are no 3D reactions defined.

I have not added either of these commands to the QRG since we may wish to
change the names to something more memorable, and we might not want to ever
allow centered molecules.  (Randomizing the positions doesn't take
measurably more time.)  We might only want viz output to report centered
positions (for convenient viewing).