Author: Rex Kerr
It is now possible to write reactions that place lots of molecules on the
surface even if there isn't room in a single grid element to hold more than
one molecule.  To do this, you need to increase VACANCY_SEARCH_DISTANCE from
its default value of zero, e.g.:
  VACANCY_SEARCH_DISTANCE=0.05
It's only useful if this distance is larger than the spacing between
adjacent effector grid elements.

Normally, if a reaction requests more grid molecules than will fit at the
location of the reaction, then the reaction is blocked.  In particular, if B
is a grid molecule, then
  B -> B+B
will always fail because there is no room for the second B.  (Note that once
I finish adding the capability for multiple effectors per grid element, this
won't be the case when B has area less than 0.5.)

If VACANCY_SEARCH_DISTANCE is greater than zero, the reaction code will
instead search for nearby free grid elements whose centroids are within
VACANCY_SEARCH_DISTANCE of the place where the reaction should occur.  If
there are enough such elements, it will pick the closest one(s) and place
the molecules there.

In particular, if you have a reaction of the form
  B -> B+B
you can use it to start with a single effector and eventually tile the
entire surface with B's.  This works regardless of whether
CENTER_MOLECULES_ON_GRID is TRUE or FALSE, although the spreading is
slightly different in each case (with TRUE it's more regular and hence
somewhat more artifact-prone).

Right now, such reactions will spread to the entire surface.  A future
modification to the reaction syntax will allow such reactions to be
restricted to certain surface classes; in this case, all products have to
fit on a surface of a compatible class (which may occupy only part of a
geometrically connected object).