Author: Rex Kerr
Added the LIST type for region releases.  This uses a new function,
insert_grid_molecule in vol_util.c, that allows placement of molecules on a
grid as long as you give it a coordinate that is nearby, and a search
distance.

Also updated documentation accordingly.  (I had mostly specified the syntax
in advance, but I forgot the overloading of the SITE_DIAMETER directive.)