Author: Rex Kerr
Reactions between grid molecules are now in.  I haven't added documentation
yet, because it doesn't really make sense, in most cases, to use the feature
with pairs of fixed molecules.  But if you really want to, you can just
define things the natural way and it will work.  Note that the rate constant
is in units of (molecules)/(um^2 * sec).

For fun, try reactions
  A->A+A    [1e5]
  B->B+B    [1e5]
  A+B->NULL [1e2]
with an effector grid density of 10^4.