Author: Rex Kerr
Fixed a bug in insert_grid_molecule (a couple of typos in the
out-of-subvolume search algorithm).

Also fixed several bugs in the checkpoint code: added a test to see if
amp->properties==NULL, which it will be if something else has reacted with
it and the scheduler hasn't hit it yet (i.e. it's not really there).  Also
test for insert_grid_molecule returning NULL (which it can if there's
nothing in range) and used a different radius.  Using the radius of a
release site is certainly not what we want to do (e.g. there may not even be
any release sites)--I've just set it to 10 for now.  Also, 3D molecules were
writing a random orient value, but expecting a value of 0, so it'd try to
add 3D molecules to surfaces.  This caused all kinds of havoc.  I've
doubly-fixed it: the molecules now look at their current properties (since
people may change a 3D molecule to a 2D one and vice versa--not advised but
it can be done!), and the 3D molecules properly set orient before being
written.

There's still something wrong--surface molecules aren't getting put back
properly--but I'm tired of debugging for now.  Not sure whether it's an
error my code or Boris' at this point.

Finally, took out comments around STEP directive in the parser.  I don't
know why this was commented out, but I needed it for testing, and it's back
now.