Author: Rex Kerr
Fixed a bug in closest_pt_point_triangle or whatever it is called.  The bug
was of the classic local-variable-is-temporary type:
  void i_do_not_actually_do_anything(struct vector3 *a,struct vector3 *b)
  {
    b = a;
  }
Alarmingly, this was actually *working* with -O3 set.  Anyway, I switched it
to memcpy.  This fixes a stack-corrupting crash bug with -O0 -ggdb3, but
doesn't fix the other problems.