Author: Rex Kerr
I've finally fixed the bugs that were causing non-identical results before
and after checkpointing.  One was intentional, actually: the code tries to
align molecule scheduling with checkpoint boundaries, which originally
seemed like a good idea but is exactly wrong for what we want to do now.
I've taken out some of that code.  I also have aligned scheduling times that
are close to an integer boundary with that boundary, so that molecules don't
jump randomly between "this iteration" and "next iteration" depending on
floating point rounding error.  And I fixed a bug that could potentially
cause a division-by-zero floating point error.  (It seemed to only crop up
with -O1 optimizations?)

One known bug remains with checkpointing: output in viz runs from 0 through the
time of checkpointing, giving constant output, instead of starting at the
time of the checkpoint.