Author: Rex Kerr
Added comments to all functions in mdlparse_util.c that I wrote entirely
myself.  There are still a number of MCell 2 functions that have no summary
comments, and a couple of hybrid functions (e.g. handle_count_request) that
contain a lot of old code and a lot of new code that do not have summary
functions.