Author: Rex Kerr (with assistance from Tom)
Modified the release pattern code.  The following changes were made.  First,
all the intervals and durations now have to be greater than zero (except for
the delay time).  Second, the default values for the intervals and release
duration is FOREVER, currently set to 1e20.  This is "forever" because it
cannot be reached during the simulation--we allow only 2^64 iterations, and
2^64 is about 1.8e19 < 1e20.  The scheduling code and default release
pattern has been updated to take advantage of this new policy, and the code
looks cleaner (to me, anyway) as a result.  Finally, we re-instituted the
old policy regarding events that happen on a "downtick", i.e. exactly at the
end of a train duration.  In MCell 2, these events did not happen, so this
is the behavior in MCell3 now also.