Author: Rex Kerr
Added garbage collection to the scheduler.  This is useful for molecules
that are scheduled infinitely far into the future and then get consumed when
they're hit by another molecule--without garbage collection, these would
remain in memory indefinitely.  The garbage collector currently is triggered
any time more than 20% of the total items in the scheduler are garbage.  It
doesn't collect garbage from the current timestep, because that stuff is
weeded out during runtime.

It also doesn't currently keep track of garbage that is removed at runtime,
so the 20% is only an estimate (the true amount of garbage could be less due
to runtime-removal).

The scheduler itself doesn't automatically handle garbage collection (or
counting of what is garbage or not)--it just provides an integer variable
for you to keep track of garbage, and a function to remove everything that
you deem to be garbage (which is returned as a linked list so you can
deallocate it).