Author: Rex Kerr
Fixed a memory leak inside releases on regions.

Also adjusted the garbage collection code to be more aware of ongoing
garbage collection during the simulation (it's still apparently only
partially aware, however--I have some logic error somewhere, but it doesn't
affect results; the GC is just called a bit too much).

Also, there is a memory leak in the NMJ model; I have to change a bunch of
code to find it, so I'm committing now so I can use CVS as my "good copy" of
the code.