Author: Rex Kerr
A good deal of code is in to allow user control of notifications.
Notification statements go in MDL inside NOTIFICATIONS {...} blocks (this
prevents accidental assignment to a variable if the user has a typo in the
name).  Notifications that work are
  DIFFUSION_CONSTANT_REPORT = OFF | BRIEF | ON
  REACTION_PROBABILITY_REPORT = OFF | ON
  RELEASE_EVENT_REPORT = OFF | ON
  FINAL_SUMMARY_REPORT = OFF | ON
  ITERATION_NUMBER_REPORT = OFF | ON | user-supplied interval
Notifications that partly work are
  PROGRESS_REPORT = OFF | ON
  ALL_NOTIFICATIONS = OFF | ON
Notifications that do not yet work are
  BOX_TRIANGULATION_REPORT = OFF | ON
  PARTITION_LOCATION_REPORT = OFF | ON
  REACTION_PROBABILITY_REPORT_THRESHOLD = user-supplied probability
  FILE_OUTPUT_REPORT = OFF | ON

Some code is in to allow control of warnings and errors.  There are lots of
places in the code where there are warnings or errors, however, and it's
going to take a while to catch them all and give them all keywords.
Warnings can be ignored (no output, code deals with it as best it can),
reported (message output, code deals), or treated as errors (message output,
code halts).  Warning statements go inside WARNINGS {...} blocks.

Warnings that work are
  MOLECULE_LIFETIME_TOO_SHORT = IGNORED | WARNING
  MOLECULE_LIFETIME_THRESHOLD = user-supplied time in iterations
  MISSED_REACTIONS = IGNORED | WARNING
  MISSED_REACTION_WARNING_THRESHOLD = user-supplied probability
Warnings that have flags planned but do not yet work are
  NEGATIVE_DIFFUSION_CONSTANT = IGNORED | WARNING | ERROR
  NEGATIVE_REACTION_PROBABILITY = IGNORED | WARNING | ERROR
  HIGH_REACTION_PROBABILITY = user-supplied probability
  CLOSE_PARTITION_SPACING = IGNORED | WARNING | ERROR
  DEGENERATE_POLYGONS = IGNORED | WARNING | ERROR
  OVERWRITTEN_OUTPUT_FILE = IGNORED | WARNING | ERROR
There are a bunch of other warnings that don't have flags yet.

None of this is in the QRG yet, since it's not finished (and it may be
sensible to remove _REPORT from all of the notifications).

(I'm committing this now because I may want to work on this code on a
different computer, plus it's far enough along to be useful.)