Author: Rex Kerr
Introduced a new boundary condition.  So far we have REFLECTIVE,
TRANSPARENT, and ABSORPTIVE.  The new one is
  CONCENTRATION my_favorite_molecule = concentration_in_molar
This works like ABSORPTIVE--that is, any molecules that hit it disappear
(they "go through" to an infinite sink on the other side).  But molecules
are also emitted as if they diffused through from an infinite sink on the
other side where the concentration is fixed at concentration_in_molar.

Internally, the stuff is referred to as a concentration clamp, so there's
various stuff added with "clamp" in the name.  For example, there's a new
ccn_clamp_data structure to keep track of all the different surface types
and molecules on which there is clamping, some new stuff in init.c to set up
this structure, and a new top-level simulation function
run_concentration_clamp that releases molecules right before run_timestep
moves them.

There's a new flag, ACT_CLAMPED, that makes new particles move according to
the correct distribution (it's different).  The ACT_INERT functionality
doesn't seem to be in, but when it does go in ACT_CLAMPED particles should
not be interacted with--if another particle hits an ACT_CLAMPED particle, it
should ignore it.  Only after the flag is off are the molecules really in
the world.

The documentation has been updated to reflect this new functionality.