Author: Rex Kerr
Finished implementing the notification and warning stuff.  Close partition
spacing is an uncaught error, so that's not in, and we don't yet overwrite
files in a clever way as far as I know, so the overwriting warning is also
not in.  Everything else should be (but is very hastily tested, since it's
getting awfully late).

Read the QRG to see what the syntax ended up being.