Tom Bartol:
  Modify all memory allocation failure messages to report useful information
in a consistent way.  Rather than saying something completely mysterious like
"memory allocation error" with no details or saying "Cannot store blah" (nice
but still a bit cryptic), messages now conform to:

  "Out of memory while creating blah"

This give the user a chance to reduce the size of their model in the
appropriate way and try again, get more RAM, or move to a bigger machine.

Please try to stick to this form in future messages.