Author: Rex Kerr
This commit contains a fully armed and operational implementation of surface
diffusion.  This has touched on a lot of parts of the code, including some
of the core scheduling routines.  Just give surface molecules a non-zero
diffusion constant and it will (ought to) work.

There is a NEW POLICY for surface reactions.  You must include orientation
marks in order to get an oriented reaction.  If you omit them, you'll have
an unoriented reaction.  This is, by default, an error, although you can set
it to a warning or whatever by using an option in the WARNINGS block that
I'll document in the next commit.

If you want an unoriented molecule, use ; or ,' or ',.  In reactions this
may not work as intended yet.