Author: Rex Kerr
Changed the policy about requiring tick marks on special surface properties;
the default is now to affect both orientations for both surface and volume
molecules.  Volume molecules can go one-way if you like however, i.e. a box
with surface property TRANSPARENT=A' will let A in but not out.