Author: Rex Kerr
Fixed a problem with microscopic reversibility for surfaces (forgot to set
the flag on the molecule properly).

To turn this on, use
  MICRO_REV_SURF_RX=TRUE
So far it doesn't seem to cause any improvement, however.

There is also a problem with the zero orientation class.  I'm not sure
what's causing it.  For now, use different orientation classes.