Author: Rex Kerr
Got halfway to adding a new count command, TRIGGER (used in place of COUNT).
It's described in the QRG and parses, but it doesn't work yet.  (I'm
committing because I want to be able to work on this on other computers.)

Also rewrote some of the reaction data output parsing code to fix a memory
leak, make column reporting generic, and handle the new TRIGGER syntax.
Also, the old code treated one-column output as a special case that didn't
get headers.  If wildcards are used, it's not safe to assume that people
will know what that one column is, so headers are now on even with only one
column.  However, this breaks xmgrace, so in the next commit I'm going to
make headers off by default unless there's a good reason to do otherwise.

Also added the ability to give custom labels to column headers.  Just do
  : "my_custom_header_title"
after the COUNT statement (or arithmetic combinations of count statements).