Author: Rex Kerr
I got sick of these silly license issues so I wrote a different
implementation of wildcard matching.  This one also allows escape strings,
so you can, if you want, search for text strings that contain * and ? and
stuff in them.  The escape character is \, as usual, so

  alpha*beta matches alphabeta, alpha77777beta, alpha*beta, etc.
  alpha\*beta matches only alpha*beta
  [([{]stuff[)\]}] matches (stuff) [stuff] {stuff} (stuff] {stuff) etc.
  c:\\Windows\\Why???\? matches c:\Windows\WhyNot?

I haven't removed the old wildcardfit function, but I suggest we try
replacing it with the new is_wildcard_match function (which is in util.c).

The new code doesn't have the raw speed of the old one, but it does more
efficiently handle constructs that match the end of a string (e.g. "*open")
and which make heavy use of character sets after wildcards (e.g.
"*?[aeiouy]*").  So it has worse best-case performance but better worst-case
performance.