Author: Rex Kerr
Created new output symbols and implemented intelligent overwriting of
reaction data output files during checkpointing.

The standard output symbol, =>, will work just like before, except that if
it runs from a checkpoint file it will look through the file if it exists
and truncate it at the point where output seems to come from a later time
than the start of the checkpoint data.  Thus, if a run fails, you can simply
rerun it and you'll get the correct data in the output files.

In addition to =>, we also now have >, +>, >>, and >>>.

> works just like the Unix redirect--it overwrites the file, always.

+> always requires a new file for the output.

>> also works like the Unix redirect--it appends to the file, always.  It
only puts headers on if it starts at the beginning of the file, however.

>>> works like >> but will always write headers, even into the middle of the
file.