Author: Rex Kerr
Added reversible arrows.  The syntax is in for catalytic arrows too, but the
code doesn't work on them yet.  I'm committing now in case I break something
while enabling catalytic arrows.

Also fixed a bug where one could write
  A + B -> C
where A and B had 3D diffusion constants but C had a 2D diffusion constant.
This is not a sensible thing to do, since there is no place to put C, and it
was causing a segfault.  The parser complains instead, now.

I may have changed or fixed something else too, but if so I forget what.