Author: Rex Kerr
I ripped out and rewrote essentially the entire count system.  It should now
be back to doing what it was supposed to be doing before--but now it is
actually doing it instead of almost, sort of doing it in a very peculiar
way.  There's only about half as many lines of code now as there was before.

There are a raft of new functions, none of which have comments.  (Those will
go in after I'm done with the counting features.)

Reactions on regions appear to be working now (essentially accidentally).

Also, there's a nifty new use of wildcards:

  SUM( COUNT["AChR.*",WORLD] )

will add up all the counts of stuff starting with "AChR.".  One can imagine
why this would be handy.  If there are lots of species, override the name:

  SUM( COUNT["AChR.*",WORLD] ) : "total_AChR"

You can use numeric constants inline:

  5*COUNT[AChR,WORLD]

and if you want trig functions and stuff, EXPRESSION works:

  EXPRESSION[SIN(PI/4)]*COUNT[AChR,WORLD]

Triggering still does not work, but will soon.  Math on wildcard expansions
is allowed when it is sensible but the implementation is currently broken,
so don't do it yet.  Empty wildcard expansions are a syntax error.  All the
front/back hits stuff should be working, but I haven't tested it all so
someone else should do this.