Author: Rex Kerr
Well, that was easy--I just forgot to supply a return value in
join_oexpr_tree.

You can do math with wildcards, as long as it is sensible math: you have to
use operations with scalars only.  So

  COUNT["A*",WORLD]-COUNT[backgroundA,WORLD]

will produce a set of counts of things starting with A, each minus the other
count.  However,

  COUNT["A*",WORLD]+COUNT["B",WORLD]

is not straightforwardly interpretable and will fail.

Incidentally the operations allowed in count statements are + - * / () and
unary -, plus the SUM( ) function.

Also, note that division by zero produces zero.  This is to allow division
by things that might be zero some of the time without having to implement an
if/else or max operation.