Author: Rex Kerr
I really screwed up the last commit in a number of different ways.  The
current stuff works a lot better, but it is still badly broken.  Don't use
this, or the previous commit.  I had a mostly-working copy for a while, but
I broke it while continuing to add triggers, and broke it even more when
reverting to an old copy to commit without broken triggers.  And now I don't
have a mostly-working copy any more.  (I should have used my secondary SVN
repository to maintain my own changes.  Whoops.)

Sorry about the problems.  I'll fix the problems this weekend.  (This commit
will conveniently allow me to take the current code on the plane.)