Authors Tom Bartol and Rex Kerr:
Rex and I have officially begun the process of Vulcan Mind Meld on the MCell3
source code.  As we proceed from day to day, we'll be adding numerous comments
and clean-ups to the code.  This is the first installment of such tweaks.

Also, added file code_notes.txt as a place to temporarily record notations and
thoughts that should eventually be added as comments or clarifications in
the actual source code.  Essentially, the notes stand as placeholders or
reminders for confusing concepts we encounter early in the Mind Meld that
we plan to clarify later on as we get deeper into the code.