Author: Rex Kerr
Counting code is changed again--this time the part that actually tracks the
change of statistics at runtime has been largely (though not completely)
rewritten.

Preliminary testing indicates that counting works in the following cases:
  World counting of volume molecules
  World counting of surface molecules
  World counting of reactions
  Region counting of volume molecules
  Region counting of reactions
The following have not been properly tested:
  Region counting of surface molecules
  Counting of hits
  Counting on objects instead of regions
  Counting of concentration
  Counting inside multiply nested or highly convoluted regions
The following are known to be broken:
  Counting of triggered reacions
  Counting of triggered volume molecules
The following are not even fully implemented:
  Counting of triggered diffusing surface molecules

The new code is slower at inserting/removing molecules but faster at keeping
track of them as they move around.  Whether this is a net win or a net loss
will depend on the details of the simulation.  (With lots of transparent
surfaces and not too many releases/reactions, it will be a win--so this
change will hopefully make adding transparent measurement regions less
noticable than it was.)

Also, updated a few messages to include or not include the file and line
numbers; any error save out-of-memory generated by the input file ought not
give MCell3 source code line numbers, as it's quite confusing when trying to
fix one's MDL (as I inadvertantly found out--"line 367 looks just fine, what
is...oh, that's a source code file!").