Author: Rex Kerr
Fixed some bugs in triggering relating to output of diffusing surface
molecules.  This now works more sensibly (you don't get two updates every
movement), although when a molecule travels out of a region you get the
position it last existed within a region, not the position at which it
crossed the border that leaves the region.