Author: Rex Kerr
Added the ability to trigger release events with reactions.

Note that this is *not tested yet*.

In the MDL file, all one has to do is give a named reaction pathway instead
of a release pattern and the release will occur there.  The LOCATION
directive is then relative to the location of the release.  If the reaction
happens at a wall, the coordinate system will also be rotated so that the
surface normal is in the +Z direction (the other coordinates will rotate as
appropriate for tipping the Z-axis directly to the surface normal).

The code works by adding a "magic_list" to the reaction pathname struct.  If
a release object refers to the reaction pathname, that object is added to
the magic list.  Then, when a reaction happens it checks the list on the
pathname, and if the release is listed there, it calls release_molecules.
Just like magic!

This isn't terribly fast, since the code has to calculate matrix transforms
on the fly.

One can extend this type of idea (which is basically just a standard event
listener model) to handle other wizardry.  Even with just this, one can
already do some pretty weird stuff.  (E.g. molecule teleportation where
molecules disappear from a volume and appear on a surface somewhere else
in the model; molecules that don't diffuse but randomly jump around inside a
region, etc..)